+++ src/main/java/com/brokurly/dto/GoodsDto.java
@@ -16,7 +16,7 @@ public class GoodsDto {
   private String name;
   private String exp;
   private int price;
-  private double dcRt;
+  private int dcRt;
   private int itemDcAmt;
   private String origin;
   private String pointFl;
